Abstracted full-text schema augmentation
